# LEAF-4123: Optimize "group designated" hotspot