This expands on earlier optimization efforts and improves performance when "group designated" records are involved. The optimization logic was largely copied from the "person designated" implementation to minimize complexity at the cost of some code duplication.

Performance improvement compared to Sprint-64-c1 when testing with ~500 "group designated" records:
```
goos: windows
goarch: amd64
pkg: LEAF/API-tester
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                           │ sprint64c1.txt │     optimize-group-selected.txt     │
                           │     sec/op     │   sec/op     vs base                │
Inbox_nonAdminActionable-8     156.62m ± 4%   96.32m ± 2%  -38.50% (p=0.000 n=10)
```

Total improvement since Sprint-60-c2:
```
goos: windows
goarch: amd64
pkg: LEAF/API-tester
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                           │ sprint60c2.txt │     optimize-group-selected.txt     │
                           │     sec/op     │   sec/op     vs base                │
Inbox_nonAdminActionable-8     426.97m ± 4%   96.32m ± 2%  -77.44% (p=0.000 n=10)
```

### Potential Impact
- Areas depending on form/query actionable data (Inbox, Report Builder)
- form/query access checks relating to "group designated" records

### Testing
- [x] You can access records that have been assigned to a designated group, and you are a member of that group, and you can take action on the record if it's on the requisite step - [experimental API test](https://github.com/mgaoVA/LEAF_test_experiments/commit/2166733fb04268569ded7ca03baf956894510496#diff-b7b1922b928b6c4ca9f0342a10188972bc7aa07897032f05faff1cb2b96f5234R59)
- [x] You cannot access records (as non-admin) that have been assigned to a designated group - [experimental API test](https://github.com/mgaoVA/LEAF_test_experiments/commit/2166733fb04268569ded7ca03baf956894510496#diff-b7b1922b928b6c4ca9f0342a10188972bc7aa07897032f05faff1cb2b96f5234R63)